Modify: Change File Structure
